Remove non-plutus-ledger types
